+++ server/controllers/source.js
@@ -1,6 +1,8 @@
 const Source = require('../models/source');
 const validateObjectId = require('./validators/objectId.validator');
 const validateSource = require('./validators/source.validator');
+const Transaction = require('../models/transaction');
+const { ObjectId } = require('mongodb');
 
 const getSources = async (req, res) => {
   const sources = await Source.find({
@@ -30,9 +32,17 @@ const createSource = async (req, res) => {
     return res.status(400).send(error.details[0].message);
   }
 
+  const existedSources = await Source.findOne({
+    $and: [{ userId: req.user.id }, { name: sourceData.name.trim() }, { type: sourceData.type.trim() } ],
+  });
+
+  if (existedSources) {
+    return res.status(403).send({message: 'This source is already exist!'});
+  }
+
   const source = await Source.create({
-    name: sourceData.name,
-    type: sourceData.type,
+    name: sourceData.name.trim(),
+    type: sourceData.type.trim(),
     userId: req.user.id,
   });
 
@@ -49,8 +59,8 @@ const updateSource = async (req, res) => {
   const source = await Source.findByIdAndUpdate(
     req.params.id,
     {
-      name: sourceData.name,
-      type: sourceData.type,
+      name: sourceData.name.trim(),
+      type: sourceData.type.trim(),
     },
     { new: true }
   );
@@ -66,6 +76,12 @@ const deleteSource = async (req, res) => {
   const { error } = validateObjectId(req.params.id);
   if (error) {
     return res.status(400).send(error.details[0].message);
+  }
+
+  const transactionExists = await Transaction.find({userId: ObjectId(req.user.id), sourceId: ObjectId(req.params.id)})
+
+  if (transactionExists && transactionExists.length) {
+    return res.status(400).send({message: "Some transactions are already associated with this source!"});
   }
 
   const source = await Source.findByIdAndDelete(req.params.id);
